worker-service/requirements.txt
# no dependencies for now
boto3
Pillow